+++ Cargo.toml
@@ -1,22 +1,3 @@
-[package]
-name = "fhe_http"
-version = "0.1.0"
-edition = "2021"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tfhe = { version = "*", features = [
-    "boolean",
-    "shortint",
-    "integer",
-    "aarch64-unix",
-] }
-serde = { version = "1.0.104", features = ["derive"] }
-serde_json = "1.0.48"
-clap = "4.5.1"
-base64 = "0.22.0"
-bincode = "1.3.3"
-
-[dev-dependencies]
-criterion = { version = "0.5.1", features = ["html_reports"] }
+members = ["fhe_http_core", "fhe_http_python"]
