+++ fhe_http_core/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "fhe_http_core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tfhe = { version = "*", features = [
+    "boolean",
+    "shortint",
+    "integer",
+    "aarch64-unix",
+] }
+serde = { version = "1.0.104", features = ["derive"] }
+serde_json = "1.0.48"
+clap = "4.5.1"
+base64 = "0.22.0"
+bincode = "1.3.3"
+
+[dev-dependencies]
+criterion = { version = "0.5.1", features = ["html_reports"] }
