+++ fhe_http_core/examples/response.txt
