+++ fhe_http_core/src/main.rs
@@ -2,7 +2,7 @@
 mod apis;
 mod configs;
 mod fhe_traits;
-mod pyo3;
+//mod pyo3;
 mod utils;
 use std::fs::File;
 use std::io::Result;
