+++ fhe_http_core/src/utils/base64.rs
