+++ fhe_http_python/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "fhe_http_python"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "fhe_http_python"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.21.2", features = ["extension-module"] }
+fhe_http_core = { path = "../fhe_http_core" }
