+++ fhe_http_python/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod apis {
+    pub mod fhe_types;
+    pub mod key_generator;
+}
